+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -9,6 +9,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 - Added support for listing all Event Grid topics in a subscription via the command `azmcp_eventgrid_topic_list`. [[#43](https://github.com/microsoft/mcp/pull/43)]
 - Added support for retrieving knowledge index schema information in Azure AI Foundry projects via the command `azmcp_foundry_knowledge_index_schema`. [[#41](https://github.com/microsoft/mcp/pull/41)]
 - Added support for listing service health events in a subscription via the command `azmcp_resourcehealth_service-health-events_list`. [[#367](https://github.com/microsoft/mcp/pull/367)]
+- Added nodepool list tool for AKS managed cluster: [[#360](https://github.com/microsoft/mcp/pull/360)]
 
 ### Breaking Changes
 
@@ -28,11 +29,18 @@ The Azure MCP Server updates automatically by default whenever a new release com
 
 #### Dependency updates
 
+<<<<<<< HEAD
 - Updated the following dependencies:
   - Azure.Identity: `1.14.0` → `1.15.0` [[#352](https://github.com/microsoft/mcp/pull/352)]
   - Azure.Identity.Broker: `1.2.0` → `1.3.0` [[#352](https://github.com/microsoft/mcp/pull/352)]
   - Microsoft.Azure.Cosmos.Aot: `0.1.1-preview.1` → `0.1.2-preview.1` [[#383](https://github.com/microsoft/mcp/pull/383)]
 - Updated the following dependencies to improve .NET Ahead-of-Time (AOT) compilation support: [[#363](https://github.com/microsoft/mcp/pull/363)]
+=======
+- Updated the following dependencies: [[#352](https://github.com/microsoft/mcp/pull/352)]
+  - Azure.Identity: `1.14.0` → `1.15.0`
+  - Azure.Identity.Broker: `1.2.0` → `1.3.0`
+- Updated the following dependencies to improve .NET Ahead-of-Time (AOT) compilation support:
+>>>>>>> 94163e4f (Update CHANGELOG for rebase)
   - Azure.ResourceManager.StorageCache: `1.3.1` → `1.3.2`
 
 ## 0.5.12 (2025-09-04)
@@ -69,7 +77,7 @@ AOT- Added a verb to the namespace name for bestpractices [[#109](https://github
 
 #### Dependency Updates
 
-- Updated the following dependencies to improve .NET Ahead-of-Time (AOT) compilation support: 
+- Updated the following dependencies to improve .NET Ahead-of-Time (AOT) compilation support:
   - Microsoft.Azure.Cosmos `3.51.0` → Microsoft.Azure.Cosmos.Aot `0.1.1-preview.1`. [[#37](https://github.com/microsoft/mcp/pull/37)]
 
 ## 0.5.8 (2025-08-21)